% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_crop_by_table.R
\name{assign_crops_by_table}
\alias{assign_crops_by_table}
\title{Assign crop by table}
\usage{
assign_crops_by_table(
  output_obj,
  field_crop_tbl,
  on_dup = c("last", "first", "error"),
  strict = TRUE
)
}
\arguments{
\item{output_obj}{A ALGR output object}

\item{field_crop_tbl}{a table connecting fields to crops}

\item{on_dup}{How to handle duplicate rows for the same field with "last", "first", "error"}

\item{strict}{If TRUE, stop on any invalid rows; if FALSE, drop invalid rows with a warning.}
}
\value{
A ALGR output object with crops assigned
}
\description{
Assign crop by table
}
\examples{
r<-terra::rast(matrix(1, nrow=50, ncol=50))
map<-establish_pac(potential_space= r,
                  cell_size=1,
                  includsion_value = 1,
                  mean_field_size = 100,
                  sd_field_size = 25,
                  distribution = "norm",
                  mean_shape_index = 1,
                  sd_shape_index = 0.3,
                  percent = 75)

assignments <- data.frame(
field = c(1, 3, 7, 10),   # field indices in map$field_list
crop  = c(2, 1, 3, 2)     # your crop ids
)

output2 <- assign_crops_by_table(map, assignments)
return_by_crop(output2, method = 1)
}
