% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_farmer_by_table.R
\name{assign_farmers_by_table}
\alias{assign_farmers_by_table}
\title{Title}
\usage{
assign_farmers_by_table(
  output_obj,
  field_farmer_tbl,
  on_dup = c("last", "first", "error"),
  strict = TRUE
)
}
\arguments{
\item{output_obj}{A ALGR output object}

\item{field_farmer_tbl}{a table connecting fields to farmers}

\item{on_dup}{How to handle duplicate rows for the same field with "last", "first", "error"}

\item{strict}{If TRUE, stop on any invalid rows; if FALSE, drop invalid rows with a warning.}
}
\value{
A ALGR output object with crops assigned
}
\description{
Title
}
\examples{
r<-terra::rast(matrix(1, nrow=50, ncol=50))
map<-establish_pac(potential_space= r,
                  cell_size=1,
                  includsion_value = 1,
                  mean_field_size = 100,
                  sd_field_size = 25,
                  distribution = "norm",
                  mean_shape_index = 1,
                  sd_shape_index = 0.3,
                  percent = 75)

set.seed(1245)

assign_tbl <- data.frame(field = c(1, 3, 7, 10), farmer = c(2, 1, 3, 2))

map2 <- assign_farmers_by_table(
 output_obj       = map,
 field_farmer_tbl = assign_tbl)




return_by_farmer(map2, method = 1)

}
