% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_dl.R
\name{establish_dl}
\alias{establish_dl}
\title{Field establishment based on a dead leave algorithm}
\usage{
establish_dl(
  potential_space,
  cell_size,
  includsion_value,
  mean_field_size,
  sd_field_size,
  distribution = "norm",
  mean_shape_index,
  sd_shape_index,
  percent
)
}
\arguments{
\item{potential_space}{a raster including a potential space category for field placement}

\item{cell_size}{cell size for output}

\item{includsion_value}{inclusion value for the potential space raster}

\item{mean_field_size}{mean field size counted by number of cells}

\item{sd_field_size}{sd field size counted by number of cells}

\item{distribution}{distribution type for field size}

\item{mean_shape_index}{mean shape index calculated by a relation between width/length of placement}

\item{sd_shape_index}{sd shape index calculated by a relation between width/length of placement}

\item{percent}{percent of the potential space to be filled with fields}
}
\value{
ALGR object
}
\description{
Field establishment based on a dead leave algorithm
}
\examples{
r<-terra::rast(matrix(1, nrow=100, ncol=100))
output <- establish_dl(potential_space = r,
                                                cell_size = 1,
                                                includsion_value = 1,
                                                mean_field_size = 50,
                                                sd_field_size = 25,
                                                distribution = "norm",
                                                mean_shape_index = .5,
                                                sd_shape_index = .1,
                                                percent = 0.95)
return_by_field(output)
}
