% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_pac.R
\name{establish_pac}
\alias{establish_pac}
\title{Establish landscape by place and conquer}
\usage{
establish_pac(
  potential_space,
  cell_size,
  includsion_value,
  additional_lim = NULL,
  mean_field_size,
  sd_field_size,
  distribution = "norm",
  mean_shape_index,
  sd_shape_index,
  percent
)
}
\arguments{
\item{potential_space}{a raster including a potential space category for field placement}

\item{cell_size}{cell size for output}

\item{includsion_value}{inclusion value for the potential space raster}

\item{additional_lim}{for use in case of inclusion of a raod raster}

\item{mean_field_size}{mean field size counted by number of cells}

\item{sd_field_size}{sd field size counted by number of cells}

\item{distribution}{either norm or lnorm}

\item{mean_shape_index}{mean shape index calculated by a relation between width/length of placement}

\item{sd_shape_index}{sd shape index calculated by a relation between width/length of placement}

\item{percent}{percent of the potential space to be filled with fields}
}
\value{
ALGR object
}
\description{
Establish landscape by place and conquer
}
\examples{
r<-terra::rast(matrix(1, nrow=50, ncol=50))
output<-establish_pac(potential_space= r,
                         cell_size=1,
                         includsion_value = 1,
                         mean_field_size = 200,
                         sd_field_size = 100,
                         distribution = "norm",
                         mean_shape_index = 3,
                         sd_shape_index = 0.3,
                         percent = 70)


return_by_field(output)
}
