% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_tes.R
\name{establish_tes}
\alias{establish_tes}
\title{Establish fields by a tesselation algorithm}
\usage{
establish_tes(
  potential_space,
  includsion_value,
  mean_field_size,
  sd_field_size,
  mean_shape_index,
  sd_shape_index
)
}
\arguments{
\item{potential_space}{a raster including a potential space category for field placement}

\item{includsion_value}{inclusion value for the potential space raster}

\item{mean_field_size}{mean field size counted by number of cells}

\item{sd_field_size}{sd field size counted by number of cells}

\item{mean_shape_index}{mean shape index calculated by a relation between width/length of placement}

\item{sd_shape_index}{sd shape index calculated by a relation between width/length of placement}
}
\value{
ALGR object
}
\description{
Establish fields by a tesselation algorithm
}
\examples{

# Create a raster with some potential space
test<-ALGR::generate_pn(50,50,1,2,3,0.01,TRUE, "land_percentage", percetange = 75)
terra::plot(test)

# Run the simulation
result <- establish_tes(
 potential_space = test,
 includsion_value = 1,
 mean_field_size = 25,
 sd_field_size = 25,
 mean_shape_index = 2,
 sd_shape_index = 2)

return_by_field(result, method = 1)

}
