% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roads.R
\name{generate_roads}
\alias{generate_roads}
\title{A function for generating artificial roads}
\usage{
generate_roads(road_length, slope)
}
\arguments{
\item{road_length}{A number}

\item{slope}{A lope raster}
}
\value{
A raster
}
\description{
A function for generating artificial roads
}
\examples{
test<-generate_pn(200,200,1,2,3,0.005,FALSE, "land_percentage", percentage = 50)
test2<-generate_roads(2000,test)
raster::plot(test2)

}
