% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{trans_1lr}
\alias{trans_1lr}
\title{A one layer transformation that allows the expansion of the potential space to
areas with submoptimal landscapes based on a co-variant probability}
\usage{
trans_1lr(rast, landcover, aggregation, arable_val = 1)
}
\arguments{
\item{rast}{A type feature raster}

\item{landcover}{A landcover raster}

\item{aggregation}{A number that defines how aggregated he potential space will be}

\item{arable_val}{the value of arable fields in your landcover raster}
}
\description{
A one layer transformation that allows the expansion of the potential space to
areas with submoptimal landscapes based on a co-variant probability
}
\examples{
set.seed(123)
original_potential_space<-generate_pn(200,200,1,2,3,0.002,TRUE,
                                               "land_percentage", percentage = 30)
corresponding_fields<-establish_pac(potential_space= original_potential_space,
                                                cell_size=1,
                                                includsion_value = 1,
                                                mean_field_size = 200,
                                                sd_field_size = 100,
                                                distribution = "norm",
                                                mean_shape_index = 3,
                                                sd_shape_index = 0.3,
                                                percent = 95)

map<-return_by_arable_land(corresponding_fields, method =2)
set.seed(123)
modified_potential_space<-generate_pn(200,200,1,2,3,0.002,TRUE,
                                               "land_percentage", percentage = 50)

result<-ALGR::trans_1lr(modified_potential_space,map,5, arable_val = 1)
par(mfrow=c(1,2))
terra::plot(original_potential_space)
terra::plot(result)





}
